[DSA] Harden top-k v1/v2 kernels against negative padded seq_lens
Negative per-row lengths (DP-padded / idle-companion rows, e.g. -4 from
GLM 5.2 MTP draft-extend metadata) are reinterpreted as ~4e9-token rows
by the unsigned device-side reads in both top-k kernels. Depending on
how the unsigned chunk arithmetic wraps, this is either an illegal
memory access (v1: any negative; v2: e.g. -1048576 on every dispatch
shape) or a silently garbage output row (v2: -1/-4), all reproduced
deterministically on B200. #30378 clamped two DSA triton producers, but
the kernels stayed exposed to every other caller (the DSv4 DP-attention
path has no clamp).

Per review feedback, handle this with signed comparisons instead of an
arithmetic clamp: negative rows take the documented trivial all-(-1)
path (a non-negative candidate index never qualifies against a negative
length). v1 keeps seq_len signed through the trivial dispatch and
naive_transform; v2 dispatches via a signed is_trivial() in both the
main and fused small-batch kernels; the plan kernel counts signed and
proves rows positive before the (unsigned) pool-routing compare, so a
full-uint32 static_cluster_threshold keeps its meaning. Zero cost:
paired timing on all six dispatch shapes is unchanged within noise.

+++ python/sglang/jit_kernel/csrc/deepseek_v4/topk_v1.cuh
@@ -53,10 +53,10 @@ SGL_DEVICE void naive_transform(
     const int32_t* __restrict__ page_table,
     int32_t* __restrict__ indices,
     int32_t* __restrict__ raw_indices,  // optional: output raw abs position indices
-    const uint32_t length,
+    const int32_t length,               // signed: negative (padded row) fills everything with -1
     const uint32_t page_bits) {
   static_assert(kTopK <= kTopKBlockSize);
-  if (const auto tx = threadIdx.x; tx < length) {
+  if (const auto tx = threadIdx.x; static_cast<int32_t>(tx) < length) {
     indices[tx] = page_to_indices(page_table, tx, page_bits);
     if (raw_indices != nullptr) {
       raw_indices[tx] = tx;
@@ -236,7 +236,11 @@ __global__ void topk_transform_kernel(const __grid_constant__ TopKParams params)
   const uint32_t work_id = blockIdx.x;
 
   /// NOTE: dangerous prefetch seq_len before PDL wait
-  const uint32_t seq_len = seq_lens[work_id];
+  // Kept signed: a negative length (DP-padded / idle-companion row) must not
+  // be reinterpreted as a ~4e9-token row (illegal memory access). The signed
+  // trivial-path comparison below routes it to naive_transform, which fills
+  // the whole row with -1 (no non-negative index qualifies).
+  const int32_t seq_len = seq_lens[work_id];
   const auto score_ptr = scores + work_id * score_stride;
   const auto page_ptr = page_table + work_id * page_table_stride;
   const auto indices_ptr = page_indices + work_id * kTopK;
@@ -244,11 +248,11 @@ __global__ void topk_transform_kernel(const __grid_constant__ TopKParams params)
 
   device::PDLWaitPrimary<kUsePDL>();
 
-  if (seq_len <= kTopK) {
+  if (seq_len <= static_cast<int32_t>(kTopK)) {
     naive_transform(score_ptr, page_ptr, indices_ptr, raw_indices_ptr, seq_len, page_bits);
   } else {
     __shared__ int32_t s_topk_indices[kTopK];
-    radix_topk(score_ptr, s_topk_indices, seq_len);
+    radix_topk(score_ptr, s_topk_indices, static_cast<uint32_t>(seq_len));
     static_assert(kTopK <= kTopKBlockSize);
     const auto tx = threadIdx.x;
     if (kTopK == kTopKBlockSize || tx < kTopK) {
